% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\docType{data}
\name{default_tags}
\alias{default_tags}
\title{Default list of safe tags}
\format{
An object of class \code{list} of length 69.
}
\usage{
default_tags
}
\description{
Inspired by https://www.npmjs.com/package/sanitize-html
}
\keyword{datasets}
