% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe.R
\docType{data}
\name{safe_tags}
\alias{safe_tags}
\title{Default list of safe tags}
\format{
An object of class \code{list} of length 69.
}
\usage{
safe_tags
}
\description{
Default list of safe tags
}
\keyword{datasets}
