% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_html}
\alias{sanitize_html}
\title{Sanitize an HTML string}
\usage{
sanitize_html(html, safe_tags = default_tags)
}
\arguments{
\item{html}{An HTML string to be sanitized}

\item{safe_tags}{the list of tags which are to be permitted during sanitization.}
}
\description{
Sanitize an HTML string
}
