% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_markdown}
\alias{sanitize_markdown}
\title{Sanitize a markdown string}
\usage{
sanitize_markdown(md, safe_tags = default_tags)
}
\arguments{
\item{md}{Markdown to be parsed and sanitized.}

\item{safe_tags}{the list of tags which are to be permitted during sanitization.}
}
\description{
Which can also contain HTML inside of it.
}
