% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_node}
\alias{sanitize_node}
\title{Sanitize an xml2 node}
\usage{
sanitize_node(node, safe_tags = default_tags)
}
\arguments{
\item{node}{An xml2 node}

\item{safe_tags}{the list of tags which are to be permitted during sanitization.}
}
\description{
Sanitize an xml2 node
}
